pyproject.toml
[project]
name = "hyperlocal"
version = "0.1.0"
description = "Hyperlocal flyer generation pipeline"
requires-python = ">=3.12"
dependencies = [
    "boto3>=1.34.0",
    "httpx>=0.27.0",
    "openai>=1.40.0",
    "psycopg[binary]>=3.2.0",
    "pydantic>=2.8.0",
    "python-dotenv>=1.0.1",
    "sqlalchemy>=2.0.30",
    "tenacity>=8.2.3",
]
